+++ lib/IRGen/AllocStackHoisting.cpp
@@ -58,8 +58,15 @@ static bool isHoistable(AllocStackInst *Inst, irgen::IRGenModule &Mod) {
     return false;
 
   // Don't hoist weakly imported (weakly linked) types.
-  auto *nominalDecl = SILTy.getASTType()->getNominalOrBoundGenericNominal();
-  if (nominalDecl->isWeakImported(Mod.getSILModule().getSwiftModule()))
+  bool foundWeaklyImported =
+      SILTy.getASTType().findIf([&Mod](CanType type) -> bool {
+        if (auto nominal = type->getNominalOrBoundGenericNominal())
+          if (nominal->isWeakImported(Mod.getSILModule().getSwiftModule())) {
+            return true;
+          }
+        return false;
+      });
+  if (foundWeaklyImported)
     return false;
 
   // Don't hoist generics with opened archetypes. We would have to hoist the
