+++ test/IRGen/Inputs/weak_import_native_helper.swift
@@ -121,3 +121,15 @@ extension ProtocolWithWeakMembers {
 
 public protocol BaseP {}
 @_weakLinked extension S : BaseP {}
+
+
+public func getVersion() -> Int {
+  // Used to return 0.
+  return 1
+}
+
+@_weakLinked public struct ResilientStruct {
+  public init() {}
+
+  public func fn(_ x: Int) {}
+}
